Move tests into packages.
